Size specifications, IE7 fix and dynamic styles
When using this polyfill I found the following issues:

* It was caching the size and position info. This isn't compatible
  with this information getting changed dynamically (for example
  through @media styles).
* Added support for specifying exact sizes.
  For example background-size: 100% 100%
* There was a extra comma that caused problems in IE7.

This is a rough patch. I basically just fixed it up for my website,
copy and pasted the final code into GitHub and submitted the patch
in hopes it is useful to someone else. If I get a chance down the
road I can try to make a proper patch (commit for each issue I am
fixing and more extensive testing to make sure I didn't break
anything). But I figured a messy patch is better than no patch if
this helps someone out and I never get a chance to clean it up.